Move old webauthn orm code from core/store/orm/orm.go to core/sessions/orm.go

Group webauthn endoints after UserController endpoints in router.go

Update go mod dependencies for webauthn duo library

Move webauthn models from model to session module (where User and Session structs were moved as well, these make sense to stay scoped together)

Move new webauthn ORM methods to sessions/orm.go - rewrite with new SQL queries pattern (INSERT as well)

Move RPID and RPOrigin to general config struct

Rename and conform webauthn migration to 0064 with +goose

+++ core/services/log/broadcaster.go
@@ -2,13 +2,13 @@ package log
 
 import (
 	"context"
+	"math/big"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"github.com/pkg/errors"
 
-	"github.com/smartcontractkit/chainlink/core/null"
+	"go.uber.org/atomic"
 	"gorm.io/gorm"
 
 	"github.com/ethereum/go-ethereum/common"
@@ -17,15 +17,16 @@ import (
 
 	"github.com/smartcontractkit/chainlink/core/internal/gethwrappers/generated"
 	"github.com/smartcontractkit/chainlink/core/logger"
+	"github.com/smartcontractkit/chainlink/core/null"
 	"github.com/smartcontractkit/chainlink/core/service"
 	"github.com/smartcontractkit/chainlink/core/services/eth"
 	httypes "github.com/smartcontractkit/chainlink/core/services/headtracker/types"
-	"github.com/smartcontractkit/chainlink/core/store/models"
 	"github.com/smartcontractkit/chainlink/core/utils"
 )
 
 //go:generate mockery --name Broadcaster --output ./mocks/ --case=underscore --structname Broadcaster --filename broadcaster.go
 //go:generate mockery --name Listener --output ./mocks/ --case=underscore --structname Listener --filename listener.go
+//go:generate mockery --name Config --output ./mocks/ --case=underscore --structname Config --filename config.go
 
 type (
 	// The Broadcaster manages log subscription requests for the Chainlink node.  Instead
@@ -61,9 +62,10 @@ type (
 	}
 
 	broadcaster struct {
-		orm       ORM
-		config    Config
-		connected *abool.AtomicBool
+		orm        ORM
+		config     Config
+		connected  *abool.AtomicBool
+		evmChainID big.Int
 
 		// a block number to start backfill from
 		backfillBlockNumber null.Int64
@@ -81,17 +83,18 @@ type (
 
 		chStop                chan struct{}
 		wgDone                sync.WaitGroup
-		trackedAddressesCount uint32
+		trackedAddressesCount atomic.Uint32
 		replayChannel         chan int64
-		highestSavedHead      *models.Head
-		lastSeenHeadNumber    int64
+		highestSavedHead      *eth.Head
+		lastSeenHeadNumber    atomic.Int64
+		logger                *logger.Logger
 	}
 
 	Config interface {
 		BlockBackfillDepth() uint64
 		BlockBackfillSkip() bool
-		EthFinalityDepth() uint
-		EthLogBackfillBatchSize() uint32
+		EvmFinalityDepth() uint32
+		EvmLogBackfillBatchSize() uint32
 	}
 
 	ListenerOpts struct {
@@ -121,15 +124,17 @@ type (
 var _ Broadcaster = (*broadcaster)(nil)
 
 // NewBroadcaster creates a new instance of the broadcaster
-func NewBroadcaster(orm ORM, ethClient eth.Client, config Config, highestSavedHead *models.Head) *broadcaster {
+func NewBroadcaster(orm ORM, ethClient eth.Client, config Config, logger *logger.Logger, highestSavedHead *eth.Head) *broadcaster {
 	chStop := make(chan struct{})
 
 	return &broadcaster{
 		orm:              orm,
 		config:           config,
+		logger:           logger,
 		connected:        abool.New(),
-		ethSubscriber:    newEthSubscriber(ethClient, config, chStop),
-		registrations:    newRegistrations(),
+		evmChainID:       *ethClient.ChainID(),
+		ethSubscriber:    newEthSubscriber(ethClient, config, logger, chStop),
+		registrations:    newRegistrations(logger, *ethClient.ChainID()),
 		logPool:          newLogPool(),
 		addSubscriber:    utils.NewMailbox(0),
 		rmSubscriber:     utils.NewMailbox(0),
@@ -154,11 +159,11 @@ func (b *broadcaster) BackfillBlockNumber() null.Int64 {
 }
 
 func (b *broadcaster) TrackedAddressesCount() uint32 {
-	return atomic.LoadUint32(&b.trackedAddressesCount)
+	return b.trackedAddressesCount.Load()
 }
 
 func (b *broadcaster) ReplayFromBlock(number int64) {
-	logger.Infof("LogBroadcaster: Replay requested from block number: %v", number)
+	b.logger.Infof("LogBroadcaster: Replay requested from block number: %v", number)
 	select {
 	case b.replayChannel <- number:
 	default:
@@ -175,6 +180,7 @@ func (b *broadcaster) Close() error {
 
 func (b *broadcaster) awaitInitialSubscribers() {
 	defer b.wgDone.Done()
+	b.logger.Debug("LogBroadcaster: Starting to await initial subscribers until all dependents are ready...")
 	for {
 		select {
 		case <-b.addSubscriber.Notify():
@@ -195,28 +201,26 @@ func (b *broadcaster) awaitInitialSubscribers() {
 
 func (b *broadcaster) Register(listener Listener, opts ListenerOpts) (unsubscribe func()) {
 	if len(opts.LogsWithTopics) == 0 {
-		logger.Fatal("LogBroadcaster: Must supply at least 1 LogsWithTopics element to Register")
+		b.logger.Fatal("LogBroadcaster: Must supply at least 1 LogsWithTopics element to Register")
 	}
 
 	reg := registration{listener, opts}
 	wasOverCapacity := b.addSubscriber.Deliver(reg)
 	if wasOverCapacity {
-		logger.Error("LogBroadcaster: Subscription mailbox is over capacity - dropped the oldest unprocessed subscription")
+		b.logger.Error("LogBroadcaster: Subscription mailbox is over capacity - dropped the oldest unprocessed subscription")
 	}
 	return func() {
 		wasOverCapacity := b.rmSubscriber.Deliver(reg)
 		if wasOverCapacity {
-			logger.Error("LogBroadcaster: Subscription removal mailbox is over capacity - dropped the oldest unprocessed removal")
+			b.logger.Error("LogBroadcaster: Subscription removal mailbox is over capacity - dropped the oldest unprocessed removal")
 		}
 	}
 }
 
-func (b *broadcaster) Connect(head *models.Head) error { return nil }
-
-func (b *broadcaster) OnNewLongestChain(ctx context.Context, head models.Head) {
+func (b *broadcaster) OnNewLongestChain(ctx context.Context, head eth.Head) {
 	wasOverCapacity := b.newHeads.Deliver(head)
 	if wasOverCapacity {
-		logger.Tracew("LogBroadcaster: Dropped the older head in the mailbox, while inserting latest (which is fine)", "latestBlockNumber", head.Number)
+		b.logger.Debugw("LogBroadcaster: TRACE: Dropped the older head in the mailbox, while inserting latest (which is fine)", "latestBlockNumber", head.Number)
 	}
 }
 
@@ -239,7 +243,7 @@ func (b *broadcaster) startResubscribeLoop() {
 
 	var chRawLogs chan types.Log
 	for {
-		logger.Debug("LogBroadcaster: Resubscribing and backfilling logs...")
+		b.logger.Debug("LogBroadcaster: Resubscribing and backfilling logs...")
 		addresses, topics := b.registrations.addressesAndTopics()
 
 		newSubscription, abort := b.ethSubscriber.createSubscription(addresses, topics)
@@ -248,7 +252,7 @@ func (b *broadcaster) startResubscribeLoop() {
 		}
 
 		if b.config.BlockBackfillSkip() && b.highestSavedHead != nil {
-			logger.Warn("LogBroadcaster: BlockBackfillSkip is set to true, preventing a deep backfill - some earlier chain events might be missed.")
+			b.logger.Warn("LogBroadcaster: BlockBackfillSkip is set to true, preventing a deep backfill - some earlier chain events might be missed.")
 			b.highestSavedHead = nil
 		}
 
@@ -270,7 +274,7 @@ func (b *broadcaster) startResubscribeLoop() {
 		}
 
 		if b.backfillBlockNumber.Valid {
-			logger.Debugw("LogBroadcaster: Using an override as a start of the backfill",
+			b.logger.Debugw("LogBroadcaster: Using an override as a start of the backfill",
 				"blockNumber", b.backfillBlockNumber.Int64,
 				"highestNumConfirmations", b.registrations.highestNumConfirmations,
 				"blockBackfillDepth", b.config.BlockBackfillDepth(),
@@ -295,11 +299,11 @@ func (b *broadcaster) startResubscribeLoop() {
 
 		b.connected.Set()
 
-		atomic.StoreUint32(&b.trackedAddressesCount, uint32(len(addresses)))
+		b.trackedAddressesCount.Store(uint32(len(addresses)))
 
 		shouldResubscribe, err := b.eventLoop(chRawLogs, subscription.Err())
 		if err != nil {
-			logger.Warnw("LogBroadcaster: Error in the event loop - will reconnect", "err", err)
+			b.logger.Warnw("LogBroadcaster: Error in the event loop - will reconnect", "err", err)
 			b.connected.UnSet()
 			continue
 		} else if !shouldResubscribe {
@@ -316,10 +320,14 @@ func (b *broadcaster) eventLoop(chRawLogs <-chan types.Log, chErr <-chan error) 
 	debounceResubscribe := time.NewTicker(1 * time.Second)
 	defer debounceResubscribe.Stop()
 
-	logger.Debug("LogBroadcaster: Starting the event loop")
+	b.logger.Debug("LogBroadcaster: Starting the event loop")
 	for {
 		select {
 		case rawLog := <-chRawLogs:
+
+			b.logger.Debugw("LogBroadcaster: Received a log",
+				"blockNumber", rawLog.BlockNumber, "blockHash", rawLog.BlockHash, "address", rawLog.Address)
+
 			b.onNewLog(rawLog)
 
 		case <-b.newHeads.Notify():
@@ -338,12 +346,12 @@ func (b *broadcaster) eventLoop(chRawLogs <-chan types.Log, chErr <-chan error) 
 
 		case blockNumber := <-b.replayChannel:
 			b.backfillBlockNumber.SetValid(blockNumber)
-			logger.Debugw("LogBroadcaster: Returning from the event loop to replay logs from specific block number", "blockNumber", blockNumber)
+			b.logger.Debugw("LogBroadcaster: Returning from the event loop to replay logs from specific block number", "blockNumber", blockNumber)
 			return true, nil
 
 		case <-debounceResubscribe.C:
 			if needsResubscribe {
-				logger.Debug("LogBroadcaster: Returning from the event loop to resubscribe")
+				b.logger.Debug("LogBroadcaster: Returning from the event loop to resubscribe")
 				return true, nil
 			}
 
@@ -366,16 +374,16 @@ func (b *broadcaster) onNewLog(log types.Log) {
 }
 
 func (b *broadcaster) onNewHeads() {
-	var latestHead *models.Head
+	var latestHead *eth.Head
 	for {
 		// We only care about the most recent head
 		item := b.newHeads.RetrieveLatestAndClear()
 		if item == nil {
 			break
 		}
-		head, ok := item.(models.Head)
+		head, ok := item.(eth.Head)
 		if !ok {
-			logger.Errorf("expected `models.Head`, got %T", item)
+			b.logger.Errorf("expected `eth.Head`, got %T", item)
 			continue
 		}
 		latestHead = &head
@@ -385,12 +393,12 @@ func (b *broadcaster) onNewHeads() {
 	// when 'b.newHeads.Notify()' receives more times that the number of items in the mailbox
 	// Some heads may be missed (which is fine for LogBroadcaster logic) but the latest one in a burst will be received
 	if latestHead != nil {
-		logger.Debugw("LogBroadcaster: Received head", "blockNumber", latestHead.Number,
+		b.logger.Debugw("LogBroadcaster: Received head", "blockNumber", latestHead.Number,
 			"blockHash", latestHead.Hash, "parentHash", latestHead.ParentHash, "chainLen", latestHead.ChainLength())
 
-		atomic.StoreInt64(&b.lastSeenHeadNumber, latestHead.Number)
+		b.lastSeenHeadNumber.Store(latestHead.Number)
 
-		keptLogsDepth := uint64(b.config.EthFinalityDepth())
+		keptLogsDepth := uint64(b.config.EvmFinalityDepth())
 		if b.registrations.highestNumConfirmations > keptLogsDepth {
 			keptLogsDepth = b.registrations.highestNumConfirmations
 		}
@@ -401,18 +409,32 @@ func (b *broadcaster) onNewHeads() {
 			keptDepth = 0
 		}
 
-		logs, minBlockNum := b.logPool.getLogsToSend(latestBlockNum)
-
-		if len(logs) > 0 {
-			broadcasts, err := b.orm.FindConsumedLogs(minBlockNum, latestBlockNum)
-			if err != nil {
-				logger.Errorf("LogBroadcaster: Failed to query for log broadcasts, %v", err)
-				return
+		// if all subscribers requested 0 confirmations, we always get and delete all logs from the pool,
+		// without comparing their block numbers to the current head's block number.
+		if b.registrations.highestNumConfirmations == 0 {
+			logs, lowest, highest := b.logPool.getAndDeleteAll()
+			if len(logs) > 0 {
+				broadcasts, err := b.orm.FindConsumedLogs(lowest, highest)
+				if err != nil {
+					b.logger.Errorf("Failed to query for log broadcasts, %v", err)
+					return
+				}
+				b.registrations.sendLogs(logs, *latestHead, broadcasts)
 			}
+		} else {
+			logs, minBlockNum := b.logPool.getLogsToSend(latestBlockNum)
 
-			b.registrations.sendLogs(logs, *latestHead, broadcasts)
+			if len(logs) > 0 {
+				broadcasts, err := b.orm.FindConsumedLogs(minBlockNum, latestBlockNum)
+				if err != nil {
+					b.logger.Errorf("LogBroadcaster: Failed to query for log broadcasts, %v", err)
+					return
+				}
+
+				b.registrations.sendLogs(logs, *latestHead, broadcasts)
+			}
+			b.logPool.deleteOlderLogs(uint64(keptDepth))
 		}
-		b.logPool.deleteOlderLogs(uint64(keptDepth))
 	}
 }
 
@@ -424,10 +446,10 @@ func (b *broadcaster) onAddSubscribers() (needsResubscribe bool) {
 		}
 		reg, ok := x.(registration)
 		if !ok {
-			logger.Errorf("expected `registration`, got %T", x)
+			b.logger.Errorf("expected `registration`, got %T", x)
 			continue
 		}
-		logger.Debugw("LogBroadcaster: Subscribing listener", "requiredBlockConfirmations", reg.opts.NumConfirmations, "address", reg.opts.Contract)
+		b.logger.Debugw("LogBroadcaster: Subscribing listener", "requiredBlockConfirmations", reg.opts.NumConfirmations, "address", reg.opts.Contract)
 		needsResub := b.registrations.addSubscriber(reg)
 		if needsResub {
 			needsResubscribe = true
@@ -444,10 +466,10 @@ func (b *broadcaster) onRmSubscribers() (needsResubscribe bool) {
 		}
 		reg, ok := x.(registration)
 		if !ok {
-			logger.Errorf("expected `registration`, got %T", x)
+			b.logger.Errorf("expected `registration`, got %T", x)
 			continue
 		}
-		logger.Debugw("LogBroadcaster: Unsubscribing listener", "requiredBlockConfirmations", reg.opts.NumConfirmations, "address", reg.opts.Contract)
+		b.logger.Debugw("LogBroadcaster: Unsubscribing listener", "requiredBlockConfirmations", reg.opts.NumConfirmations, "address", reg.opts.Contract)
 		needsResub := b.registrations.removeSubscriber(reg)
 		if needsResub {
 			needsResubscribe = true
@@ -489,14 +511,14 @@ func (b *broadcaster) appendLogChannel(ch1, ch2 <-chan types.Log) chan types.Log
 }
 
 func (b *broadcaster) maybeWarnOnLargeBlockNumberDifference(logBlockNumber int64) {
-	lastSeenHeadNumber := atomic.LoadInt64(&b.lastSeenHeadNumber)
+	lastSeenHeadNumber := b.lastSeenHeadNumber.Load()
 	diff := logBlockNumber - lastSeenHeadNumber
 	if diff < 0 {
 		diff = -diff
 	}
 
 	if lastSeenHeadNumber > 0 && diff > 1000 {
-		logger.Warnw("LogBroadcaster: Detected a large block number difference between a log and recently seen head. "+
+		b.logger.Warnw("LogBroadcaster: Detected a large block number difference between a log and recently seen head. "+
 			"This may indicate a problem with data received from the chain or major network delays.",
 			"lastSeenHeadNumber", lastSeenHeadNumber, "logBlockNumber", logBlockNumber, "diff", diff)
 	}
@@ -541,10 +563,9 @@ func (n *NullBroadcaster) AwaitDependents() <-chan struct{} {
 	close(ch)
 	return ch
 }
-func (n *NullBroadcaster) DependentReady()                                {}
-func (n *NullBroadcaster) Start() error                                   { return nil }
-func (n *NullBroadcaster) Close() error                                   { return nil }
-func (n *NullBroadcaster) Healthy() error                                 { return nil }
-func (n *NullBroadcaster) Ready() error                                   { return nil }
-func (n *NullBroadcaster) Connect(*models.Head) error                     { return nil }
-func (n *NullBroadcaster) OnNewLongestChain(context.Context, models.Head) {}
+func (n *NullBroadcaster) DependentReady()                             {}
+func (n *NullBroadcaster) Start() error                                { return nil }
+func (n *NullBroadcaster) Close() error                                { return nil }
+func (n *NullBroadcaster) Healthy() error                              { return nil }
+func (n *NullBroadcaster) Ready() error                                { return nil }
+func (n *NullBroadcaster) OnNewLongestChain(context.Context, eth.Head) {}
